[tomarkdown] Integrate `siteconfig' and refactor.
Integrate SiteConfig class in tomarkdown. The initial code is in
HtmlFormatter::__init__() and the functions to add chapters and
sections are being called while TOC is being generated.
This is the best choice because it will ensure that page ordering
is same as in Table of Contents.

Other changes include a clean-up which should not change
functionality, but if anything broken is found later, there is a
high chance that it comes from this commit.

* tomarkdown.py: Implemented changes detailed above.