[docwriter] Change docwriter to a 'package'.
Docwriter is now a python 'package' as suggested in

  http://lists.nongnu.org/archive/html/freetype-devel/2018-07/msg00224.html

There are multiple changes involved in this process:

* docwriter and its files now live in the 'docwriter/' directory.

* Entry-point 'docwriter.py' is now '__main__.py' because this is the
file run by python when `python -m docwriter` is invoked.

* New file '__init__.py' introduced. This makes 'docwriter' a package.

* All docwriter modules and tests now import as `from docwriter import {...}`.
This is because 'docwriter' is now a package.

* New file: setup.py defines package information and requirements.

This is WIP.

NOTE: Package name is currently 'docwritertest' to avoid accidentally
publishing with name 'docwriter'.