Use MkDocs 1.0, drop support for older versions.
This commit adds support for MkDocs, effectively dropping support for
any previous version. This is because of the multiple
backward-incompatible changes introduced in 1.0. The full changelog is
available at

  https://www.mkdocs.org/about/release-notes/#version-10-2018-08-03

A few changes important to docwriter:

* Static site pages are now generated directly as `.html' pages in the
`site/' directory when `use_directory_urls' is set to `false'. This is
relevant because docwriter uses this setting.

* `pages' config is now replaced with `nav'. Although docwriter does not
need to hide any page from the navigation, this change may be useful at
a later stage.

* Linking to pages outside the site in the navigation is now possible.
This allows for adding links to other websites/pages outside of the
documentation.

* docs/reference/.gitignore: Add `site/' folder.

* docwriter/siteconfig.py: Change `pages' config to `nav'.

* docwriter/tomarkdown.py: Change relative page paths.

* docwriter/utils.py (file_exists): Exit on error.

* requirements.txt: Upgrade requirements, drop support for Mkdocs < 1.0

* tests/tomarkdown.py (test_make_section_url): Change expected URL.