* requirements.txt: Use specific versions.
Changing to specific version number rather than using the `>=` notation
allows docwriter to be compatible in the long term. An external bot
(dependabot) will inform of new versions so that they can be tested
before the numbers are bumped up.

Another reason this is being done is libraries like `mkdocs` and
`mkdocs-material` are in active development. This means the possibility
of deprecation and removal of certain features important to docwriter.