Add regression tests for docwriter.
Docwriter will now be tested with the 'freetype2' repository. The
tox environment 'regression' is used, which runs the script
'tests/freetype.sh', that clones 'freetype2' and configures it.
The make targets 'refdoc' and 'refdoc-venv' are tested, and the
docs are also built with 'builddir' != 'srcdir'.

Currently the test is allowed to fail (this means that the Travis
build will be successful even if the regression test fails), but
this can be changed later, so that it becomes necessary for regression
tests to pass for the build to be successful.

Note that the testenv 'regression' is not added to [envlist] in
'tox.ini' because we don't want it to run by default when 'tox'
command is invoked locally. This is a long test with many
requirements, so we prefer to run it only on Travis.

* .travis.yml (matrix): Add regression test job to matrix.
(fast_finish): Mark the build successful as soon as all required
tests pass.
(allow_failures): Allow regression build to fail without affecting
the build status.

* CHANGELOG: Updated.

* README.md: Add more information for running tests, and general
formatting.

* tests/freetype.sh: Script for regression tests. New file.

* tox.ini (envlist): Formatting.
(testenv:regression): New test environment.